+++ easychat/lib/src/widgets/chat.bubble.dart
@@ -104,7 +104,7 @@ class ChatBubble extends StatelessWidget {
         // margin: const EdgeInsets.symmetric(vertical: 4),
         margin: const EdgeInsets.symmetric(vertical: 8),
         child: UserBlocked(
-            otherUid: message.uid!,
+            otherUid: message.uid,
             builder: (blocked) {
               if (blocked) {
                 return Row(
@@ -182,35 +182,17 @@ class ChatBubble extends StatelessWidget {
                   if (message.uid != myUid) ...[
                     Padding(
                       padding: const EdgeInsets.only(top: 4.0),
-
-                      /// TODO: don't listen to the whole user data. Only listen user display name and photo url.
                       child: UserDoc.sync(
-                        uid: message.uid!,
-                        builder: (user) {
-                          if (user == null) {
-                            return Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(20),
-                                color: Theme.of(context)
-                                    .colorScheme
-                                    .primaryContainer,
+                        uid: message.uid,
+                        builder: (user) => user == null
+                            ? const ChatAvatarLoader()
+                            : GestureDetector(
+                                behavior: HitTestBehavior.opaque,
+                                onTap: () => UserService.instance
+                                    .showPublicProfileScreen(context,
+                                        user: user),
+                                child: UserAvatar(user: user),
                               ),
-                              width: 48,
-                              height: 48,
-                              child: const CircularProgressIndicator(),
-                            );
-                          }
-                          return GestureDetector(
-                            behavior: HitTestBehavior.opaque,
-                            onTap: () {
-                              UserService.instance.showPublicProfileScreen(
-                                context,
-                                user: user,
-                              );
-                            },
-                            child: UserAvatar(user: user),
-                          );
-                        },
                       ),
                     ),
                     const SizedBox(width: 8),
@@ -226,15 +208,12 @@ class ChatBubble extends StatelessWidget {
                           : CrossAxisAlignment.end,
                       children: [
                         if (message.uid != myUid) ...[
-                          /// TODO: don't listen to the whole user data. Only listen user display name and photo url.
-                          UserDoc.sync(
-                            uid: message.uid!,
-                            builder: (user) {
-                              if (user == null ||
-                                  user.displayName.trim().isEmpty) {
-                                return const Text("...");
-                              }
-                              return Text(user.displayName);
+                          UserField(
+                            uid: message.uid,
+                            initialData: message.displayName,
+                            field: 'displayName',
+                            builder: (v, r) {
+                              return Text(v ?? '...');
                             },
                           ),
                           const SizedBox(width: 8),
@@ -435,49 +414,5 @@ class ChatBubble extends StatelessWidget {
       );
     }
     return const SizedBox.shrink();
-  }
-}
-
-class ChatProtocolBubble extends StatelessWidget {
-  const ChatProtocolBubble({
-    super.key,
-    required this.message,
-  });
-
-  final ChatMessage message;
-
-  @override
-  Widget build(BuildContext context) {
-    final protocol = message.protocol!;
-    String text = protocol.t;
-    if (protocol == ChatProtocol.join) {
-      text = protocol.tr(
-        args: {
-          'displayName': message.displayName,
-        },
-      );
-    }
-    return Padding(
-      padding: const EdgeInsets.all(12),
-      child: Row(
-        children: [
-          const Expanded(
-            child: Divider(
-              thickness: 1,
-              endIndent: 16,
-            ),
-          ),
-          Text(
-            text,
-          ),
-          const Expanded(
-            child: Divider(
-              thickness: 1,
-              indent: 12,
-            ),
-          ),
-        ],
-      ),
-    );
   }
 }